refactor(react-divider): refactor Divider component and remove useDividerBase hook
