+++ change/@fluentui-react-divider-3b499241-7e2d-4a58-8f8b-a7b1de4ea065.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "refactor: remove useDividerBase hook",
+  "packageName": "@fluentui/react-divider",
+  "email": "dmytrokirpa@microsoft.com",
+  "dependentChangeType": "patch"
+}
