+++ packages/react-components/react-divider/library/src/Divider.ts
@@ -1,15 +1,8 @@
-export type {
-  DividerBaseProps,
-  DividerProps,
-  DividerSlots,
-  DividerBaseState,
-  DividerState,
-} from './components/Divider/index';
+export type { DividerBaseProps, DividerProps, DividerSlots, DividerState } from './components/Divider/index';
 export {
   Divider,
   dividerClassNames,
   renderDivider_unstable,
   useDividerStyles_unstable,
   useDivider_unstable,
-  useDividerBase_unstable,
 } from './components/Divider/index';
