+++ packages/react-components/react-divider/library/src/components/Divider/Divider.types.ts
@@ -43,9 +43,5 @@ export type DividerProps = ComponentProps<Partial<DividerSlots>> & {
   vertical?: boolean;
 };
 
-export type DividerBaseProps = Omit<DividerProps, 'alignContent' | 'appearance' | 'inset' | 'vertical'>;
-
 export type DividerState = ComponentState<DividerSlots> &
   Required<Pick<DividerProps, 'alignContent' | 'appearance' | 'inset' | 'vertical'>>;
-
-export type DividerBaseState = Omit<DividerState, 'alignContent' | 'appearance' | 'inset' | 'vertical'>;
