+++ packages/react-components/react-divider/library/src/components/Divider/index.ts
@@ -2,5 +2,4 @@ export { Divider } from './Divider';
 export type { DividerBaseProps, DividerProps, DividerSlots, DividerBaseState, DividerState } from './Divider.types';
 export { renderDivider_unstable } from './renderDivider';
 export { useDivider_unstable } from './useDivider';
-export { useDividerBase_unstable } from './useDividerBase';
 export { dividerClassNames, useDividerStyles_unstable } from './useDividerStyles.styles';
