+++ packages/react-components/react-divider/library/src/index.ts
@@ -6,7 +6,3 @@ export {
   useDivider_unstable,
 } from './Divider';
 export type { DividerProps, DividerSlots, DividerState } from './Divider';
-
-// Experimental APIs - will be uncommented in the experimental release branch
-// export { useDividerBase_unstable } from './Divider';
-// export type { DividerBaseProps, DividerBaseState } from './Divider';
